Application: integrate RunLoop and Windows Message Loop
Use SPI to construct a custom event source for the Windows message loop
and integrate it into an instance of the main RunLoop.  This allows us
to use Foundation to pump the Windows message loop as well as the
dispatch queues.  With this, it is now possible to use dispatch to run
UI processing asynchronously without having to manually pump the
dispatch queue.