Keep inline with master fork
* Update azure-pipelines.yml

Fetch modulemaps from `main` branch.

* UI: remove unnecessary variable in table view processing

We can use the sentinel `-1` for the index to indicate that the item
should be added to the end.  This avoids the need for special counting
and will become more important once header and footer cells are
inserted.

* Improve Application main loop

Handle RunLoop (CFRunLoop) timers/sources and/or Dispatch main queue tasks in application main loop

* UI: support creating an empty view

This enables the creation of an empty view which can be used to build
general views that are not control views.

* Update uses of GetLastError() to be called within the win32 init of Error (#188)

Replace direct usage of `GetLastError` in favour of the wrapped `Error` type which gives us provenance information as well as a human readable form of the error message.

* Revert "Improve Application main loop"

This reverts commit 8706996e99610a0cf462501756a666986daca3ac.

With this change UICatalog immediately exits upon launch rather than
continuing to pump the main loop.  The application launch completes
successfully as the key window is rendered, but then terminates shortly
thereafter.  Revert the change until the mainloop can be pumped.

Co-authored-by: Saleem Abdulrasool <compnerd@compnerd.org>
Co-authored-by: Yury Vovk <vovk.yury@me.com>